refactor: split metadata tests into a separate class
